pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/airflow_aws"]

[project]
name = "airflow-aws-demo"
version = "0.1.0"
description = "Modern Apache Airflow 2.10.5 with AWS Integration - Testing ground for Airflow ideas"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Airflow Demo Team", email = "demo@example.com" }
]
dependencies = [
    "apache-airflow==2.10.5",
    "apache-airflow-providers-amazon>=8.30.0",
    "apache-airflow-providers-http>=4.13.0",
    "boto3>=1.35.0",
    "botocore>=1.35.0",
    "anthropic>=0.39.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.2",
    "typing-extensions>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest-env>=1.1.0",
    "pytest-xdist>=3.6.0",
    "pytest-timeout>=2.2.0",
    "moto[all]>=5.0.0",
    "ruff>=0.8.0",
    "black>=24.10.0",
    "mypy>=1.13.0",
    "isort>=5.13.0",
    "pre-commit>=4.0.0",
    "ipython>=8.29.0",
    "ipdb>=0.13.13",
    "faker>=30.0.0",
    "freezegun>=1.5.0",
]

test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest-env>=1.1.0",
    "pytest-xdist>=3.6.0",
    "moto[all]>=5.0.0",
    "faker>=30.0.0",
    "freezegun>=1.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src",
    "--cov=plugins",
    "--cov=dags",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml:coverage.xml",
    "--cov-fail-under=85",
]
markers = [
    "unit: Unit tests - test individual components in isolation",
    "integration: Integration tests - test AWS service integrations with moto",
    "e2e: End-to-end tests - test complete DAG executions",
    "slow: Slow running tests",
    "aws: Tests requiring AWS mocking",
    "s3: S3-specific tests",
    "secrets: Secrets Manager tests",
    "bedrock: Bedrock/AI tests",
    "dag: DAG execution tests",
]
env = [
    "AIRFLOW_HOME=/tmp/airflow_test",
    "AIRFLOW__CORE__DAGS_FOLDER=dags",
    "AIRFLOW__CORE__PLUGINS_FOLDER=plugins",
    "AIRFLOW__CORE__LOAD_EXAMPLES=False",
    "AIRFLOW__CORE__UNIT_TEST_MODE=True",
    "AWS_DEFAULT_REGION=us-east-1",
    "AWS_ACCESS_KEY_ID=testing",
    "AWS_SECRET_ACCESS_KEY=testing",
]

[tool.coverage.run]
source = ["src", "plugins", "dags"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
precision = 2
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "plugins", "dags", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ANN", # flake8-annotations
    "S",   # flake8-bandit (security)
    "T20", # flake8-print
    "SIM", # flake8-simplify
]

ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "ANN101", # missing type annotation for self
    "ANN102", # missing type annotation for cls
    "ANN401", # dynamically typed expressions (Any)
    "S104",   # possible binding to all interfaces
    "S105",   # possible hardcoded password
    "S106",   # possible hardcoded password
    "T201",   # print statement (allowed in Airflow DAGs)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]  # Allow unused imports and import *
"tests/**/*.py" = ["S101", "ANN", "T201"]  # Allow assert, skip annotations, allow print
"dags/**/*.py" = ["T201"]  # Allow print in DAGs

[tool.ruff.lint.isort]
known-first-party = ["airflow_aws", "plugins"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "airflow.*",
    "moto.*",
    "boto3.*",
    "botocore.*",
    "anthropic.*",
]
ignore_missing_imports = true

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*", "*/.venv/*"]
known_first_party = ["airflow_aws", "plugins"]
known_third_party = [
    "airflow",
    "boto3",
    "botocore",
    "anthropic",
    "pydantic",
    "pytest",
    "moto",
]
